docs: move generate_schema_csv back to docs dir
